Use uint16_t for port instead of int everywhere
git-svn-id: https://svn.code.sf.net/p/plexil/code/trunk@458 d2be67b0-814c-0410-949f-d215b6b29573

+++ interfaces/Sockets/base/ClientSocket.cpp
@@ -30,7 +30,7 @@
 #include "SocketException.h"
 
 
-ClientSocket::ClientSocket ( std::string host, int port )
+ClientSocket::ClientSocket ( std::string host, uint16_t port )
 {
    if ( ! Socket::create() )
    {
